Add Pycycle Command-Line-Tool
Code with circular imports in python often raises an obscure Exception (ImportError: Cannot import name X) which tells nothing about the cycle.

I have come across this problem several times, and depending on the project size it becomes problematic to identify where the problem is. This tool shows you the chain of the imports that are causing the problem, after that is up to developer to fix it e.g. move the imports to the bottom, or move the imports inside some function.