Adds auto_ml, automated ML for engineers
[auto_ml](https://github.com/ClimbsRocks/auto_ml) democratizes access to machine learning to any engineer or analyst who can write Python. Many standard classification and regression problems follow largely the same set of best practices. Rewriting several hundred lines of largely duplicative code each time is a pain to maintain, and slows down iteration speeds. 

auto_ml is tuned for production, both in prediction speeds, flexibility of inputs, and serializing the trained model to disk. It is the automated machine learning platform that we use for our engineers at a large tech startup, and we're just starting to publicize it a bit and make it more widely available to the community. 

Let me know if you have any questions!