Added Smores
Smores leverages pyparsing, jinja and marshmallow to provide a system in which you are able to easily define a hierarchical data structure for your user-created templates.  Marshmallow provides data serialization and jinja provides the templating engine itself.  Pyparsing is used to expose a simpler tag syntax for end users.  Smores comes equipped with a 'tab to complete' style autocomplete method that also provides validation for the passed tag fragment.