Added quandl-metabase
Quandl is a reputed source for financial, economic datasets. The datasets in Quandl can be visualized with the tools which they provide in their site. This quickstart project, build with Python makes use of Metabase library to visualize the datasets from Quandl.
Let me know if you need more details.