Added yattag to libraries for working with http
Made use of yattag for a personal project where BeautifulSoup just didn't cut it to put simply.

Allows you full control of the HTML language with python commands. I made use of this in my project to take the output from a range of commands and combined this with Bootstrap for an aesthetically pleasing output. For this reason, I feel it should be added to this list