Add Raylib-Python-CFFI
Raylib and this subsequent project is a simple but effective way to learn games programming or continue your own projects with less distractions. Bindings are offered as both an exact copy of the C API and a more pythonistic API. Genuinely surprised to not see any Raylib bindings here yet but this one has impressed me the most.