Add hyperparameter
hyperparameter is a configuration framework designed for data scientists and machine learning engineers. It works in a more pythonic way:  the users write their code in python and add a decorator to the functions or classes; `hyperparameter` will map the keyword arguments to configuration keys and use a `with` context to manage the configurations.