Add Quix Streams (stream processing w Kafka)
This is a powerful commercial library which is now being open-sourced. Was developed at McLaren for processing huge sensor data streams transmitted from F1 racing cars, so useful for anyone working with sensor data or any other high-frequency time-series data,

- [https://pypi.org/project/quixstreams/](https://pypi.org/project/quixstreams/)
- [https://github.com/quixio/quix-streams](https://github.com/quixio/quix-streams)

Was originally started as a companion to the [Quix.io](http://quix.io/) platform, but can now be used as a standalone library with any Kafka installation.