Add Roundup Issue Tracker to Misc section
Hello:

Please consider adding the Roundup Issue Tracker to your list. Roundup is an issue-tracking system that boasts a user-friendly interface and easy installation process. It offers a range of interfaces, including command-line, web, REST, XML-RPC, and e-mail, making it a versatile solution for issue tracking. 

Roundup is highly customizable, allowing users to tailor the system to their specific needs and preferences.