:triangular_flag_on_post: an extra accuracy README.md
"ORM" may not encompass all database mapping tools, and the new phrasing provides clarity for users who might be unfamiliar with the term. But I believe the term "ORM" is so popular, especially around newcomers that dropping that from the title might be a bad idea. the Object Relational Mapper has an emphasis on the "Relational" side.