add mininterface
# What is this Python project?

Mininterface is an argparse alternative and drop-in replacement that exposes configuration options to the CLI, but also to GUI and TUI. This allows your script seamlessly to run everywhere, on a desktop, web, or a headless machine. It's a dialog toolkit that lets the developer focus on their program's logic, not on how users interact with it.

# What's the difference between this Python project and similar ones?

No other project enables a single script to work simultaneously as a CLI tool, desktop app, and terminal application. Typically, you'd have to choose one or build each interface separately.
What sets Mininterface apart is that it provides helpful output in both the CLI and IDEs through type hints. It requires just one line to enable, with zero learning curve for developers.

--

Anyone who agrees with this pull request could submit an Approve review to it.